Report container's CPU utilization, rather than node's.
So far, we've been using node_exporter's metrics, which reported
metrics per node. We switch to using cadvisor's per container metrics,
which measure provider performance in isolation.

So far, we've been reporting CPU utilization as a percentage of total
CPU resources. 100% CPU utilization on an 8-CPU machine meant all
cores were fully utilized. We switch to reporting CPU utilization as a
percentage of one CPU. Full utilization of an 8-CPU machine is now
reported as 800%.

Signed-off-by: Cem Mergenci <cmergenci@gmail.com>