Allow large runners to be configured by an input
Some of the reusable workflow jobs in here refer to large runners which exist within the upbound GitHub organisation. These can't be invoked by forks:

> Called workflows cannot be queued onto self-hosted runners across organisations/enterprises. Failed to queue this job. Labels: 'e2-standard-8 , linux'.

If we make this into an input then callers can customise it themselves and still continue to use the reusable workflows to publish their providers.